Takss, Projects, login adjustments, Portal/App, Cookie token

+++ assets/Functions/HTTPMethods.ts
@@ -1,10 +1,25 @@
+import Cookies from "js-cookie";
 import { DynamicallyFilledObject } from "../interfaces/DefaultTypes";
+
+async function notTokenizedpost<T>(
+  path: string,
+  body: DynamicallyFilledObject<string> | Array<any>
+): Promise<T> {
+  return await fetch(`/api${path}`, {
+    method: "POST",
+    body: JSON.stringify(body),
+    headers: {
+      "Content-Type": "application/json",
+    },
+  }).then(processResponse);
+}
 
 async function post<T>(
   path: string,
-  body: DynamicallyFilledObject<string> | Array<any>,
-  token?: string
+  body: DynamicallyFilledObject<string> | Array<any>
 ): Promise<T> {
+  let token = Cookies.get("token");
+  if (!token) return;
   return await fetch(`/api${path}`, {
     method: "POST",
     body: JSON.stringify(body),
@@ -17,9 +32,9 @@ async function post<T>(
 
 async function put<T>(
   path: string,
-  body: DynamicallyFilledObject<any> | Array<any>,
-  token?: string
+  body: DynamicallyFilledObject<any> | Array<any>
 ): Promise<T> {
+  let token = Cookies.get("token");
   return await fetch(`/api${path}`, {
     method: "PUT",
     body: JSON.stringify(body),
@@ -30,19 +45,18 @@ async function put<T>(
   }).then(processResponse);
 }
 
-async function fetchObject<T>(token: string, path: string): Promise<T> {
+async function fetchObject<T>(path: string): Promise<T> {
+  let token = Cookies.get("token");
   return await fetch(`/api${path}`, {
     headers: {
-      Authorization: `Bearer ${token}`,
+      Authorization: token ? `Bearer ${token}` : "",
     },
   }).then(processResponse);
 }
 
-async function fetchAll<T>(token: string, path: string): Promise<Array<T>> {
-  return await fetchObject<Array<T>>(token, `${path}`);
-}
-
-async function sendDelete<T>(token: string, path: string): Promise<T> {
+async function sendDelete<T>(path: string): Promise<T> {
+  let token = Cookies.get("token");
+  if (!token) return;
   return await fetch(`/api${path}`, {
     method: "DELETE",
     headers: {
@@ -64,6 +78,6 @@ export const http_methods = {
   post,
   put,
   fetch: fetchObject,
-  fetchAll,
   delete: sendDelete,
+  notTokenizedpost,
 };
